backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
aiohttp>=3.10.0
pydantic>=2.9.0
pyyaml>=6.0.2
redis>=5.2.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
openai>=1.55.0
httpx>=0.28.0
